feat: move components and hooks to shared layer
