+++ src/modules/posts/ui/hooks/usePostForm.ts
@@ -4,7 +4,7 @@ import { PostRepository } from '@/src/modules/posts/domain/PostRepository';
 import { PostDataCreate } from '@/src/modules/posts/domain/PostDataCreate';
 import { Post } from '@/src/modules/posts/domain/Post';
 import { createPost } from '@/src/modules/posts/application/create/createPost';
-import { FormStatus } from '@/src/hooks/useFormData';
+import { FormStatus } from '@/src/shared/hooks/useFormData';
 
 export function usePostForm(
 	userRepository: UserRepository,
