+++ src/pages/_app.tsx
@@ -1,7 +1,7 @@
 import '@/src/styles/globals.css';
 import type { AppProps } from 'next/app';
-import { NavBar } from '@/src/components/NavBar';
-import { Footer } from '@/src/components/Footer';
+import { NavBar } from '@/src/shared/components/NavBar';
+import { Footer } from '@/src/shared/components/Footer';
 
 export default function App({ Component, pageProps }: AppProps) {
 	return (
