+++ src/pages/create/index.tsx
@@ -2,7 +2,7 @@
 import React, { useState, useEffect } from 'react';
 import { createApiPostRepository } from '@/src/modules/posts/infra/ApiPostRepository';
 import { createApiUserRepository } from '@/src/modules/users/infra/ApiUserRepository';
-import { FormStatus, useFormData } from '@/src/hooks/useFormData';
+import { FormStatus, useFormData } from '@/src/shared/hooks/useFormData';
 import { Post } from '@/src/modules/posts/domain/Post';
 import { PostCard } from '@/src/modules/posts/ui/components/PostCard';
 import { usePostForm } from '@/src/modules/posts/ui/hooks/usePostForm';
