+++ src/shared/components/NavBar.tsx
